+++ src/main/java/study/commerceservice/domain/order/Order.java
@@ -1,9 +1,8 @@
-package study.domain.order;
+package study.commerceservice.domain.order;
 
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
