+++ src/main/java/study/commerceservice/domain/order/PaymentType.java
@@ -1,4 +1,4 @@
-package study.domain.order;
+package study.commerceservice.domain.order;
 
 public enum PaymentType {
     CREDIT_CARD
