+++ src/main/java/study/commerceservice/domain/order/Receiver.java
@@ -1,4 +1,4 @@
-package study.domain.order;
+package study.commerceservice.domain.order;
 
 import javax.persistence.Embeddable;
