+++ src/main/java/study/commerceservice/repository/OrderRepository.java
@@ -1,7 +1,7 @@
-package study.repository;
+package study.commerceservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import study.domain.order.Order;
+import study.commerceservice.domain.order.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
 }
